refactor: modularize formula engine
Split the monolithic `src/utils/formula.ts` into a `formula/` directory containing focused modules (`math.ts`, `date.ts`, `stateful.ts`, `evaluate.ts`, `parser.ts`, `compile.ts`, `wholeColumn.ts`, and `types.ts`). The original `src/utils/formula.ts` acts as a barrel file preserving the public API.
